- ILogger incorporated for ProgressLogger & UserLogger

+++ SheetLink/Model/LookupField.cs
@@ -16,7 +16,7 @@ namespace PNCA_SheetLink.SheetLink.Model
         public string ParameterType { get; set; } = string.Empty;
         public string UnitType { get; set; } = string.Empty;
         public string ForgeTypeId { get; set; } = string.Empty;
-        public Dictionary<string, int> ElementElementIdPairs { get; set; } = new Dictionary<string, int>();        
+        public Dictionary<string, long> ElementElementIdPairs { get; set; } = new Dictionary<string, long>();        
         public int FieldIndex { get; set; }
         public Parameter ParameterElement { get; set; }
         public ElementId SelectedElementId { get; set; }
